requirements.txt
# Core Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# API Clients
requests==2.31.0
httpx==0.25.1
ShopifyAPI==12.0.0
aiohttp==3.9.1

# AI & ML
openai==1.3.0
anthropic==0.7.8
langchain==0.0.335
langchain-openai==0.0.2

# Image Processing
Pillow==10.1.0
requests-toolbelt==1.0.0

# Video Generation (optional - can install later if needed)
# moviepy==1.0.3

# Task Scheduling
celery==5.3.4
redis==5.0.1

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tenacity==8.2.3

